+++ src/KeyVault/KeyVault/Commands/Secret/BackupAzureKeyVaultSecret.cs
@@ -16,6 +16,7 @@ using System;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.KeyVault.Models;
+using Microsoft.Azure.Commands.KeyVault.Models.Secret;
 using Microsoft.Azure.Commands.KeyVault.Properties;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 
@@ -35,6 +36,7 @@ namespace Microsoft.Azure.Commands.KeyVault
 
         private const string BySecretNameParameterSet = "BySecretName";
         private const string BySecretObjectParameterSet = "BySecret";
+        private const string BySecretUriParameterSet = "BySecretUri";
 
         #endregion
 
@@ -61,6 +63,17 @@ namespace Microsoft.Azure.Commands.KeyVault
         [ValidateNotNullOrEmpty]
         [Alias( Constants.SecretName )]
         public string Name { get; set; }
+
+        /// <summary>
+        /// KeyVault Secret ID (uri of the secret)
+        /// </summary>
+        [Parameter(Mandatory = true,
+           Position = 0,
+           ParameterSetName = BySecretUriParameterSet,
+           HelpMessage = "The URI of the KeyVault Secret.")]
+        [Alias("SecretId")]
+        [ValidateNotNullOrEmpty]
+        public string Id { get; set; }
 
         /// <summary>
         /// The secret object to be backed up.
@@ -103,6 +116,14 @@ namespace Microsoft.Azure.Commands.KeyVault
             {
                 Name = InputObject.Name;
                 VaultName = InputObject.VaultName;
+            }
+
+            if (ParameterSetName == BySecretUriParameterSet)
+            {
+                SecretUriComponents splitUri = new SecretUriComponents(Id);
+
+                VaultName = splitUri.VaultName;
+                Name = splitUri.SecretName;
             }
 
             if ( ShouldProcess( Name, Properties.Resources.BackupSecret ) )