+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,6 +19,15 @@
 -->
 
 ## Upcoming Release
+* Upgraded Azure.Core to 1.44.1 and Azure.Identity to 1.13.0
+
+## Version 4.0.0
+* [Breaking Change] Removed alias `Resolve-Error` for the cmdlet `Resolve-AzError`.
+* Updated the `Get-AzAccessToken` breaking change warning message.
+* Added Long Running Operation Support for Invoke-AzRest command.
+
+## Version 3.0.5
+* Fixed the issue that `Export-AzSshConfig` and `Enter-AzVM` from Az.Ssh are not able to use when WAM is enabled.
 * Added breaking change preannouncement for the removal of alias `Resolve-Error`. #26189
 * Integrated new detection library to expand the scope of secrets.