+++ src/AnalysisServices/AnalysisServices/ChangeLog.md
@@ -18,9 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Migrate Analysis SDK to generated SDK
+
+## Version 1.1.6
+* Migrated AnalysisServices SDK to generated SDK
     - Removed "Microsoft.Azure.Management.Analysis" Version "2.0.4" PackageReference
-    - Added AnalysisService.Management.Sdk ProjectReference
+    - Added AnalysisServices.Management.Sdk ProjectReference
 
 ## Version 1.1.5
 * Removed the outdated deps.json file.