+++ src/ConnectedMachine/ConnectedMachine/ChangeLog.md
@@ -18,6 +18,14 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 1.1.0
+* Updated preview version api of HybridCompute to 2024-07-31
+
+## Version 1.0.0
+* General availability for module Az.ConnectedMachine
+
+## Version 0.10.0
 * Updated stable version api of HybridCompute to 2024-07-10
 
 ## Version 0.9.0